Copying //UE4/Orion-Staging to //UE4/Main (Origin: //Orion/Dev-General @ 2904087)
==========================
MAJOR FEATURES + CHANGES
==========================

#lockdown Nick.Penwarden

Change 2903938 on 2016/03/10 by Frank.Gigliotti

	Added an instance ID to FAnimMontageInstance

	#CodeReview Laurent.Delayen
	#RB Laurent.Delayen
	#Tests PIE

Change 2903745 on 2016/03/10 by Wes.Hunt

	Update Oodle TPS
	#rb none
	#tests none
	#codereview:john.pollard

Change 2903689 on 2016/03/10 by Uriel.Doyon

	New "LogHeroMaterials" console command, displaying the current state of materials and  textures on the character hero.
	#rb marcus.wasmer
	#codereview marcus.wassmer
	#tests editor, playing PC games, trying the new command

Change 2903669 on 2016/03/10 by Aaron.McLeran

	OR-17180 Make stat soundcues and stat soundwaves NOT display zero volume sounds

	- Change only effects debug stat commands for audio guys

	#rb none
	#tests played paragon with new debug stat commands, confirms doesn't show zero-volume sounds

Change 2903625 on 2016/03/10 by John.Pollard

	XB1 Oodle SDK

	#rb none
	#tests none
	#codereview Jeff.Campeau

Change 2903577 on 2016/03/10 by Ben.Marsh

	Remaking latest build scripts from //UE4/Main @ 2900980.

Change 2903560 on 2016/03/10 by Ben.Marsh

	Initial version of BuildGraph scripts - used to create build processes in UE4 which can be run locally or in parallel across a build farm (assuming synchronization and resource allocation implemented by a separate system). Intended to supersede GUBP.

	Build graphs are declared using an XML script using syntax similar to MSBuild, ANT or NAnt, and consist of the following components:

	* Tasks: Building blocks which can be executed as part of the build process. Many predefined tasks are provided (<Cook>, <Compile>, <Copy>, <Stage>, <Log>, <PakFile>, etc...), and additional tasks may be added be declaring classes derived from AutomationTool.CustomTask in other UAT modules.
	* Nodes: A named sequence of tasks which is executed to produce outputs. Nodes may have input dependencies on other nodes before they can be executed. Declared with the <Node> element in scripts.
	* Agent Groups: A set of nodes nodes which is executed on the same machine if running as part of a build system. Has no effect when building locally. Declared with the <Group> element in scripts.
	* Triggers: Container for groups which should only be executed when explicitly triggered (using the -Trigger=<Name> or -SkipTriggers command line argument). Declared with the <Trigger> element in scripts.
	* Notifiers: Specifies email recipients for failures in one or more nodes, whether they should receive notifications on warnings, and so on.

	Properties can be passed in to a script on the command line, or set procedurally with the <Property Name="Foo" Value="Bar"/> syntax. Properties referenced with the $(Property Name) notation are valid within all strings, and will be expanded as macros when the script is read. If a property name is not set explicitly, it defaults to the contents of an environment variable with the same name.

	Local properties, which only affect the scope of the containing XML element (node, group, etc...) are declared with the <Local Name="Foo" Value="Bar"/> element, and will override a similarly named global property for the local property's scope.

	Any elements can be conditionally defined via the "If" attribute, and are largely identical to MSBuild conditions. Literals in conditions may be quoted with single (') or double (") quotes, or an unquoted sequence of letters, digits and underscore characters. All literals are considered identical regardless of how they are declared, and are considered case-insensitive for comparisons (so true equals 'True', equals "TRUE"). Available operators are "==", "!=", "And", "Or", "!", "(...)", "Exists(...)" and "HasTrailingSlash(...)". A full grammar is written up in Condition.cs.

	File manipulation is done using wildcards and tags. Any attribute that accepts a list of files may consist of: a Perforce-style wildcard (matching any number of "...", "*" and "?" patterns in any location), a full path name, or a reference to a tagged collection of files, denoted by prefixing with a '#' character. Files may be added to a tag set using the <Tag> Task, which also allows performing set union/difference style operations. Each node can declare multiple outputs in the form of a list of named tags, which other nodes can then depend on.

	Build graphs may be executed in parallel as part build system. To do so, the initial graph configuration is generated by running with the -Export=<Filename> argument (producing a JSON file listing the nodes and dependencies to execute). Each participating agent should be synced to the same changelist, and UAT should be re-run with the appropriate -Node=<Name> argument. Outputs from different nodes are transferred between agents via shared storage, typically a network share, the path to which can be specified on the command line using the -SharedStorageDir=<Path> argument. Note that the allocation of machines, and coordination between them, is assumed to be managed by an external system.

	A schema for the known set of tasks can be generated by running UAT with the "-Schema=<FileName>" option. Generating a schema and referencing it from a BuildGraph script allows Visual Studio to validate and auto-complete elements as you type.

	#rb none
	#codereview Marc.Audy, Wes.Hunt, Matthew.Griffin, Richard.Fawcett
	#tests local only so far, but not part of any build process yet

Change 2903539 on 2016/03/10 by John.Pollard

	Improve replay playback debugging of character movement

	#rb none
	#tests replays

Change 2903526 on 2016/03/10 by Ben.Marsh

	Remake changes from //UE4/Main without integration history, to add support for BuildGraph tasks.

	#rb none
	#tests none

Change 2903512 on 2016/03/10 by Dan.Youhon

	Modify minimum Duration values for JumpForce and MoveToForce ability tasks so that having minimum Duration values doesn't trigger check()s

	#rb None
	#tests Compiles

Change 2903474 on 2016/03/10 by Marc.Audy

	Fix crash if ChildActor is null
	#rb None
	#tests None

Change 2903314 on 2016/03/10 by Marc.Audy

	Fix ParentComponent not being persisted and fixup content that was saved in the window it was broken
	#rb James.Golding
	#tests Selection of child actors works as expected
	#jira UE-28201

Change 2903298 on 2016/03/10 by Simon.Tovey

	Disabling the trails optimization.

	#tests none
	#rb none

	#codereview Olaf.Piesche

Change 2903124 on 2016/03/10 by Robert.Manuszewski

	Small refactor to pak signing to help with exe protection

	#rb none
	#tests none

[CL 2907678 by Andrew Grant in Main branch]

+++ Engine/Source/Runtime/PakFile/Public/IPlatformFilePak.h
@@ -3,6 +3,24 @@
 #pragma once
 
 PAKFILE_API DECLARE_LOG_CATEGORY_EXTERN(LogPakFile, Log, All);
+DECLARE_FLOAT_ACCUMULATOR_STAT_EXTERN(TEXT("Total pak file read time"), STAT_PakFile_Read, STATGROUP_PakFile, PAKFILE_API);
+
+DECLARE_DWORD_ACCUMULATOR_STAT_EXTERN(TEXT("Num open pak file handles"), STAT_PakFile_NumOpenHandles, STATGROUP_PakFile, PAKFILE_API);
+
+// Whether to use CRC (1) for chunk hashes, or SHA1 (0)
+#define PAKFILE_USE_CRC_FOR_CHUNK_HASHES 1
+#if PAKFILE_USE_CRC_FOR_CHUNK_HASHES
+static const int64 GPakFileChunkHashSize = 4;
+#else
+static const int64 GPakFileChunkHashSize = 20;
+#endif
+
+#define PAKFILE_TRACK_SECURITY_EXCLUDED_FILES 0
+
+/**
+* Compute a hash for a given block of data, as used for pak chunk signing
+*/
+void PAKFILE_API ComputePakChunkHash(const uint8* InData, const int64 InDataSize, uint8* OutHash);
 
 /**
  * Struct which holds pak file info (version, index offset, hash value).
@@ -703,6 +721,7 @@ public:
 		, ReadPos(0)
 		, Reader(InPakFile, InPakEntry, InPakReader)
 	{
+		INC_DWORD_STAT(STAT_PakFile_NumOpenHandles);
 	}
 
 	/**
@@ -714,6 +733,8 @@ public:
 		{
 			delete Reader.PakReader;
 		}
+
+		DEC_DWORD_STAT(STAT_PakFile_NumOpenHandles);
 	}
 
 	//~ Begin IFileHandle Interface
@@ -736,6 +757,8 @@ public:
 	}
 	virtual bool Read(uint8* Destination, int64 BytesToRead) override
 	{
+		SCOPE_SECONDS_ACCUMULATOR(STAT_PakFile_Read);
+
 		// Check that the file header is OK
 		if (!Reader.PakEntry.Verified)
 		{
@@ -806,6 +829,12 @@ class PAKFILE_API FPakPlatformFile : public IPlatformFile
 	bool bSigned;
 	/** Synchronization object for accessing the list of currently mounted pak files. */
 	FCriticalSection PakListCritical;
+	/** true if we aren't allowing loose files. */
+	bool bSecurityEnabled;
+#if PAKFILE_TRACK_SECURITY_EXCLUDED_FILES
+	/** List of files already excluded by the pak security system */
+	TSet<FName> SecurityExcludedFiles;
+#endif
 
 	/**
 	 * Gets mounted pak files
@@ -887,6 +916,11 @@ class PAKFILE_API FPakPlatformFile : public IPlatformFile
 	 * @param OutPakFiles List of all found pak files
 	 */
 	static void FindAllPakFiles(IPlatformFile* LowLevelFile, const TArray<FString>& PakFolders, TArray<FString>& OutPakFiles);
+
+	/**
+	 * When security is enabled, determine if this filename can be looked for in the lower level file system
+	 */
+	bool IsFilenameAllowed(const FString& InFilename);
 
 public:
 
@@ -993,7 +1027,11 @@ public:
 			return true;
 		}
 		// File has not been found in any of the pak files, continue looking in inner platform file.
-		bool Result = LowerLevel->FileExists(Filename);
+		bool Result = false;
+		if (IsFilenameAllowed(Filename))
+		{
+			Result = LowerLevel->FileExists(Filename);
+		}
 		return Result;
 	}
 
@@ -1006,7 +1044,11 @@ public:
 			return FileEntry->CompressionMethod != COMPRESS_None ? FileEntry->UncompressedSize : FileEntry->Size;
 		}
 		// First look for the file in the user dir.
-		int64 Result = LowerLevel->FileSize(Filename);
+		int64 Result = 0;
+		if (IsFilenameAllowed(Filename))
+		{
+			Result = LowerLevel->FileSize(Filename);
+		}
 		return Result;
 	}
 
@@ -1018,7 +1060,11 @@ public:
 			return false;
 		}
 		// The file does not exist in pak files, try LowerLevel->
-		bool Result = LowerLevel->DeleteFile(Filename);
+		bool Result = false;
+		if (IsFilenameAllowed(Filename))
+		{
+			Result = LowerLevel->DeleteFile(Filename);
+		}
 		return Result;
 	}
 
@@ -1030,7 +1076,11 @@ public:
 			return true;
 		}
 		// The file does not exist in pak files, try LowerLevel->
-		bool Result = LowerLevel->IsReadOnly(Filename);
+		bool Result = false;
+		if (IsFilenameAllowed(Filename))
+		{
+			Result = LowerLevel->IsReadOnly(Filename);
+		}
 		return Result;
 	}
 
@@ -1042,7 +1092,11 @@ public:
 			return false;
 		}
 		// Files not in pak are allowed to be moved.
-		bool Result = LowerLevel->MoveFile(To, From);
+		bool Result = false;
+		if (IsFilenameAllowed(From))
+		{
+			Result = LowerLevel->MoveFile(To, From);
+		}
 		return Result;
 	}
 
@@ -1055,7 +1109,11 @@ public:
 			return bNewReadOnlyValue;
 		}
 		// Try lower level
-		bool Result = LowerLevel->SetReadOnly(Filename, bNewReadOnlyValue);
+		bool Result = bNewReadOnlyValue;
+		if (IsFilenameAllowed(Filename))
+		{
+			Result = LowerLevel->SetReadOnly(Filename, bNewReadOnlyValue);
+		}
 		return Result;
 	}
 
@@ -1068,7 +1126,11 @@ public:
 			return PakFile->GetTimestamp();
 		}
 		// Fall back to lower level.
-		FDateTime Result = LowerLevel->GetTimeStamp(Filename);
+		FDateTime Result = FDateTime::MinValue();
+		if (IsFilenameAllowed(Filename))
+		{
+			Result = LowerLevel->GetTimeStamp(Filename);
+		}
 		return Result;
 	}
 
@@ -1089,7 +1151,15 @@ public:
 		else
 		{
 			// Fall back to lower level.
-			LowerLevel->GetTimeStampPair(FilenameA, FilenameB, OutTimeStampA, OutTimeStampB);
+			if (IsFilenameAllowed(FilenameA) && IsFilenameAllowed(FilenameB))
+			{
+				LowerLevel->GetTimeStampPair(FilenameA, FilenameB, OutTimeStampA, OutTimeStampB);
+			}
+			else
+			{
+				OutTimeStampA = FDateTime::MinValue();
+				OutTimeStampB = FDateTime::MinValue();
+			}
 		}
 	}
 
@@ -1098,7 +1168,10 @@ public:
 		// No modifications allowed on files from pak (although we could theoretically allow this one).
 		if (FindFileInPakFiles(Filename) == NULL)
 		{
-			LowerLevel->SetTimeStamp(Filename, DateTime);
+			if (IsFilenameAllowed(Filename))
+			{
+				LowerLevel->SetTimeStamp(Filename, DateTime);
+			}
 		}
 	}
 
@@ -1111,7 +1184,11 @@ public:
 			return PakFile->GetTimestamp();
 		}
 		// Fall back to lower level.
-		FDateTime Result = LowerLevel->GetAccessTimeStamp(Filename);
+		FDateTime Result = false;
+		if (IsFilenameAllowed(Filename))
+		{
+			Result = LowerLevel->GetAccessTimeStamp(Filename);
+		}
 		return Result;
 	}
 
@@ -1134,7 +1211,14 @@ public:
 		}
 
 		// Fall back to lower level.
-		return LowerLevel->GetFilenameOnDisk(Filename);
+		if (IsFilenameAllowed(Filename))
+		{
+			return LowerLevel->GetFilenameOnDisk(Filename);
+		}
+		else
+		{
+			return Filename;
+		}
 	}
 
 	virtual IFileHandle* OpenRead(const TCHAR* Filename, bool bAllowWrite = false) override;
@@ -1142,12 +1226,16 @@ public:
 	virtual IFileHandle* OpenWrite(const TCHAR* Filename, bool bAppend = false, bool bAllowRead = false) override
 	{
 		// No modifications allowed on pak files.
-		if (FindFileInPakFiles(Filename) != NULL)
+		if (FindFileInPakFiles(Filename) != nullptr)
 		{
-			return NULL;
+			return nullptr;
 		}
 		// Use lower level to handle writing.
-		return LowerLevel->OpenWrite(Filename, bAppend, bAllowRead);
+		if (IsFilenameAllowed(Filename))
+		{
+			return LowerLevel->OpenWrite(Filename, bAppend, bAllowRead);
+		}
+		return nullptr;
 	}
 
 	virtual bool DirectoryExists(const TCHAR* Directory) override
